requirements.txt
matplotlib==3.10.5
pandas==2.3.1
typing-extensions==4.12.2
psutil==6.0.0
GPUtil==1.4.0
numpy==1.26.4
pytorchcv==0.0.67
wandb==0.17.5
tensorboard==2.17.1
tqdm==4.66.4
gdown==5.2.0
torchmetrics==1.4.1 
dill==0.3.8
packaging==23.2
qpsolvers[open_source_solvers]==4.3.3
librosa==0.10.1
transformers==4.55.2
wandb==0.21.1